+++ src/FunFair.CodeAnalysis.Tests/ProhibitedMethodsDiagnosticsAnalyzerTests.cs
@@ -266,6 +266,77 @@ namespace FunFair.CodeAnalysis.Tests
         }
 
         [Fact]
+        public void QueryArbitrarySqlAsyncIsBanned()
+        {
+            const string test = @"
+    using System;
+
+    namespace FunFair.Common.Data
+    {
+         public interface ISqlServerDatabase
+         {
+                Task QueryArbitrarySqlAsync(string sql);
+         }
+    }
+
+    namespace ConsoleApplication1
+    {
+        class TypeName
+        {
+            void Test(FunFair.Common.Data.ISqlServerDatabase sqlServerDatabase, string sql)
+            {
+                sqlServerDatabase.QueryArbitrarySqlAsync(sql)
+            }
+        }
+    }";
+            DiagnosticResult expected = new DiagnosticResult
+                                        {
+                                            Id = "FFS0007",
+                                            Message = @"Only use ISqlServerDatabase.QueryArbitrarySqlAsync in integration tests",
+                                            Severity = DiagnosticSeverity.Error,
+                                            Locations = new[] {new DiagnosticResultLocation(path: "Test0.cs", line: 18, column: 17)}
+                                        };
+
+            this.VerifyCSharpDiagnostic(test, expected);
+        }
+
+        [Fact]
+        public void ExecuteArbitrarySqlAsyncIsBanned()
+        {
+            const string test = @"
+    using System;
+
+    namespace FunFair.Common.Data
+    {
+         public interface ISqlServerDatabase
+         {
+                Task ExecuteArbitrarySqlAsync(string sql);
+         }
+    }
+
+    namespace ConsoleApplication1
+    {
+
+        class TypeName
+        {
+            void Test(FunFair.Common.Data.ISqlServerDatabase sqlServerDatabase, string sql)
+            {
+                sqlServerDatabase.ExecuteArbitrarySqlAsync(sql);
+            }
+        }
+    }";
+            DiagnosticResult expected = new DiagnosticResult
+                                        {
+                                            Id = "FFS0006",
+                                            Message = @"Only use ISqlServerDatabase.ExecuteArbitrarySqlAsync in integration tests",
+                                            Severity = DiagnosticSeverity.Error,
+                                            Locations = new[] {new DiagnosticResultLocation(path: "Test0.cs", line: 19, column: 17)}
+                                        };
+
+            this.VerifyCSharpDiagnostic(test, expected);
+        }
+
+        [Fact]
         public void NoErrorsReported()
         {
             const string test = @"";