+++ src/FunFair.CodeAnalysis/ProhibitedMethodsDiagnosticsAnalyzer.cs
@@ -48,7 +48,12 @@ namespace FunFair.CodeAnalysis
                                       title: @"Avoid use of inline SQL statements",
                                       message: "Only use ISqlServerDatabase.ExecuteArbitrarySqlAsync in integration tests",
                                       sourceClass: "FunFair.Common.Data.ISqlServerDatabase",
-                                      bannedMethod: "ExecuteArbitrarySqlAsync")
+                                      bannedMethod: "ExecuteArbitrarySqlAsync"),
+            new ProhibitedMethodsSpec(Rules.RuleDontUseArbitrarySqlForQueries,
+                                      title: @"Avoid use of inline SQL statements",
+                                      message: "Only use ISqlServerDatabase.QueryArbitrarySqlAsync in integration tests",
+                                      sourceClass: "FunFair.Common.Data.ISqlServerDatabase",
+                                      bannedMethod: "QueryArbitrarySqlAsync")
         };
 
         /// <inheritdoc />